+++ aom/cmd/ls.go
@@ -24,7 +24,7 @@ aom class ls --module "MyModule" --baseUrl "https://apiomat.yourcompany.com/yamb
 
 		classes, err := client.GetClasses(module)
 		if err != nil {
-			panic(err)
+			logFatal(err, "An error occurred during fetching the classes of module %v", module)
 		}
 
 		fmt.Printf("Classes of module %v:\n", module)